split request preparations from connection
