--- app/app/src/main/java/com/ar/enbaldeapp/services/NoBody.java
@@ -1,14 +1,0 @@
-package com.ar.enbaldeapp.services;
-
-import java.net.HttpURLConnection;
-
-public class NoBody implements IRequester {
-    @Override
-    public void sendRequestTo(HttpURLConnection connection) {
-    }
-
-    @Override
-    public String getBoundary() {
-        return "";
-    }
-}
