+++ app/app/src/main/java/com/ar/enbaldeapp/services/requesters/GetRequester.java
@@ -1,0 +1,24 @@
+package com.ar.enbaldeapp.services.requesters;
+
+import com.ar.enbaldeapp.services.IResponseWrapper;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+
+public class GetRequester implements IRequester {
+    private final IResponseWrapper wrapper;
+
+    public GetRequester(IResponseWrapper wrapper) {
+        this.wrapper = wrapper;
+    }
+
+    @Override
+    public void sendRequestTo(HttpURLConnection connection) throws IOException {
+        connection.setRequestMethod("GET");
+    }
+
+    @Override
+    public String preprocessResponse(String response) {
+        return wrapper.preprocessResponse(response);
+    }
+}
