+++ app/app/src/main/java/com/ar/enbaldeapp/services/requesters/IRequester.java
@@ -1,9 +1,9 @@
-package com.ar.enbaldeapp.services;
+package com.ar.enbaldeapp.services.requesters;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
 
 public interface IRequester {
     void sendRequestTo(HttpURLConnection connection) throws IOException;
-    String getBoundary();
+    String preprocessResponse(String response);
 }
